requirements.txt
astrapy~=0.6.0
fastapi~=0.99.1
langchain>=0.0.339
openai~=1.3.0
pandas~=2.0.3
pydantic~=1.10.10
python-dotenv>=1.0.0
tiktoken~=0.4.0
tqdm~=4.65.0
uvicorn~=0.22.0
